requirements.txt
beautifulsoup4==4.11.1
ChatterBot==1.0.4
fake_useragent==0.1.11
htbuilder==0.6.0
imutils==0.5.4
keras==2.7.0
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.4
pandas_profiling==2.6.0
Pillow==9.1.1
plotly==5.3.1
pyecharts==1.9.1
pytz==2021.1
requests==2.25.1
scikit_learn==1.1.1
seaborn==0.11.1
SQLAlchemy==1.2.19
streamlit==0.89.0
streamlit_webrtc==0.41.0
tensorflow==2.7.0
opencv-python-headless==4.6.0.66
jinja2<3.1.0
xlrd==1.2.0
openpyxl==3.0.9
